% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{analysis.cumulant}
\alias{analysis.cumulant}
\title{cumulant analysis of particle sizes}
\usage{
analysis.cumulant(dls, channels = c(1), lin.fit.limit = -2, start = 5)
}
\arguments{
\item{dls}{a list of dls data, e.g. from read.ALV}

\item{channels}{which channel to fit a single or an array of values (default 1)}

\item{lin.fit.limit}{the logarithm above this value will be
taken (-1 --> 1/e) (default -2)}

\item{start}{integer  starting index in the correlation array}
}
\value{
a list containing the fitted curves and extracted information
           lin.fit   linear fitting parameters returned by lm
           P.lin     polydispersity from linear fit
           P         polydispersity from exponential fit
           lin.fit.indx  index of points used in linear fit
           K.lin     Cumulant coefficients of linear fit
           K         cumulant coefficients from exponential fit
           fit       fit result of exponential fit (optim)
           par       returned fit parameters of exponential fit
           D         diffusion coefficien from exponential fit
           Rh        hydrodynamic (equivalent) radius from D
           D.lin.avg average diffusion coefficient from linear fit
           Rh.lin    hydrodynamic radius from D.lin.avg
           lin.min   the minimum of the correlation function if it did not go to 0
           y.lin.fitted  fitted correlation points with linear fit
           y.fitted      fitted correlation points with exponential fit
           tau           time delay array
}
\description{
cumulant analysis of particle sizes
}
\details{
perform a cummulant analysis on the dls$correlation data.
This should be an array with first column the tau values in milliseconds,
further columns are correlator channels (if there is more than one)
The analysis assumes the correlation being in the form of:
beta * (1 + exp( - 2 Gamma*(t - K2/2 t^2 + K3/6 t^3 + ...))

Gamma = q^2 D, where D = kB T/(6 pi eta R)
and polydispersity as P = K2/K1^2,

There are two fitting steps. First a linearized version of the equation
comparing log(g - 1) ~ t (g is the autocorrelation data)
the second doing the actual exponential fit. This latter is better for
not distorting the statistics of the y-data (one can open a long argument about
this, since the correlation is already a derived information from intensity values).
The linear fit is ln(g-1) = 2*(K0 - Gamma*tau + K2/2*tau^2-K3/(2*3)*tau^3)

The linear fit is limited to be > lin.fit.limit, the index of the used points
being returned as lin.fit.indx.
}
