% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{analysis.sizes}
\alias{analysis.sizes}
\title{Size distribution estimation from correlation data}
\usage{
analysis.sizes(
  dls,
  Rh.array = seq.log(0.00866264, 8662.64, len = 50),
  g2.channel = c(2),
  tau.min = -1,
  tau.max = -1,
  start.parm = NULL,
  weight.max = 0.66,
  alpha.array = seq.log(50, 0.005, len = 50),
  factor = 0.99,
  inherit = TRUE,
  plot.fit = TRUE
)
}
\arguments{
\item{dls}{the dls list containing the correlation data as well as
parameters, such as scattering vector, angle, temperature...}

\item{Rh.array}{an array of radii in nanometers, nm
Select the radius distribution carefully. Too low walues will bias
the plateau. The large range may have to go to up to non-physical
ranges, e.g. more than 50 microns.
Default array is taken from the ALV software}

\item{g2.channel}{string, name of the channel in the correlation array}

\item{tau.min}{float, start fitting from this delay value (ms)}

\item{tau.max}{float, fit up to this delay time (ms)
if they are set to -1, then the [3:length(g2)] indices are taken
(ALV standard values)}

\item{start.parm}{array, the distribution array to start with. By default it is estimated
using the exponential decay of the g1 function}

\item{weight.max}{float, the maximal weight allowed
It may influence the convergence, because g2 values close
to zero may kick it up errorneously}

\item{alpha.array}{an array of alpha values to be scanned through. It is good to have
it logarithmically spaced}

\item{factor}{float, a new error is suitable new result if it is less than
factor*old error. If the alphas are increasing then less or
equal is taken.}

\item{inherit}{Boolean, if TRUE, feed the best fit to the start values of
the next alpha fit}

\item{plot.fit}{Boolean, if set, plot the correlation, fit and residuals
in a split plot}
}
\value{
a list describing all possible type of results, such as
                     the resulted fit errors for various alpha values,
                     the size distribution, the corresponding Gamma, D and Rh arrays,
                     and the error parts

     end.fit     the best fit object
             The following are taken over the DLS list input:
     filename    dls$filename, filename of the original data
     time        dls$time, the timestamp of the measurement
     temperature temperature
     eta         viscosity
     q           scattering vector
     tau         delay time array of the correlation function
     weight      the weight array provided
             Fit results
     fitted      fitted correlation data
     residuals   error between correlation and fitted results
     g2          original correlation array
     g2.norm     normalization factor of correlation data
     g2.bg       constant background of correlation data
     D.array     diffusion coefficients corresponding to the size array
     G.array     exponent (1/time) array corresponding to the size array
     Rh.array    user requested size array (hydrodynamic radii)
     dist        the amplitudes of the distributed for each Rh value
     dist.norm   the normalized amplitudes of the distributed for each Rh value
                 normalization is such that the maximum is 1.
     N.dist      the number distribution amplitudes
     V.dist      the volume weighted distribution amplitudes
         For each alpha value the optimum was calculated
     alphas      alpha.array
     errs        the fit error for each alpha value
     x.norm      the curvature part of err
     f.norm      the fit errors part of err
     end.alpha   the selected best alpha value
}
\description{
Size distribution estimation from correlation data
}
\details{
A regularized fitting to construct the correlation function in dynamic
 light scattering as a sum of exponentials.
 The method implemented here is between the CONTIN and a direct regularized fit.
 It fits the g(2)-1 function (squared) not its square root as in the literature, avoiding
 the extra squaring of the result.

 The intensity autocorrelation is background corrected (constant subtracted) and
 normalized to 1 before fitting. The fitted result is then scaled back to the original
 data.
 For every requested alpha value the fit is performed using the Rh array.
 The exponents are Gamma = 2*q**2*(kB T)/(6 pi eta Rh),
 for every tau delay in tau.array the exponent goes:
     func.array <- exp(outer(tau, -G.array, "*"))
 from which the error is (alpha.parm is one of the alpha.array)
 sum(
     (weights*(func.array%*%x - g2))^2) + sum((alpha.parm*Om%*%x)[-c(1,N)]^2
     )

 Using boxed constrains, the parameters are limited to the [0,1] closed interval.
 The best alpha is searched in the alpha.array, to provide the minimum of the
 fit error f.norm.
 If inherit is TRUE, the previous result will be fed as starting parameters
 If start.parm is NULL, a zero array is used for starting parameter,
 with one point set to 0.5 for the radius corresponding to where the correlation
 function falls to 1/e it amplitude.

 Running with decreasing alpha values, we get the most details of the fit which still
 improved the fit by factor to the previous best.
 (High alpha means smoothness dominates the distribution.)
}
\references{
The algorithm was inspired by:
     the size distribution of microgel suspensions' in The Journal of Chemical Physics
     vol. 142, 234905 (2015)
and the original CONTIN papers:
     S. W. Provencher, 'CONTIN: A general purpose constrained regularization program
     for inverting noisy linear algebric and integral equations' in
     Computer Physics Communications vol. 27, 229 - 242 (1982)
and
     S. W. Provencher, 'A constrained regularization method for inverting data represented
     by linear algebric or integral equations' in
     Computer Physics Communications vol.: 27, 213 - 227 (1982)
}
