% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{correlation.limit.sizes}
\alias{correlation.limit.sizes}
\title{recalculate correlation function with limited size window}
\usage{
correlation.limit.sizes(fit, Rh.min = NULL, Rh.max = NULL)
}
\arguments{
\item{fit}{a size fit list object from analysis.sizes}

\item{Rh.min}{beginning of R-range to be kept, default NULL}

\item{Rh.max}{end of R-range to be kept, default NULL}
}
\value{
alist containing
         g2.orig     original correlation data
         g2.calc     new corelation function
         correlation array with tau and the residual correlation
         q           q from dls
         eta         viscosity from dls
         temperature from dls
         D.array     diffusion coefficients for Rh.array
         G.array     exponents for the Rh.array
         Rh.array    the Rh.array
         dist        weights of the size distribution
}
\description{
recalculate correlation function with limited size window
}
\details{
take a size fit object (list resulted by analysis.sizes)
plot the size distribution and activate a selector
so the user can select a range to be kept

If Rh.min and / or Rh.max are specified, skip the selection plot,
use them to define the range

Calculate the correlation function of the selection,
 and subtract from the original correlation function
plot what is remaining and return it as a table

Required fields in fit:
g2 the correlation function
tau the delay times
g2.norm --> normalization factor for g2
g2.bg --> background constant in the g2 data
during fit the (g2 - g2.bg)/g2.norm was fit
}
