% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{g.2}
\alias{g.2}
\title{g.2 correlation from delay times, exponents and a weight distribution}
\usage{
g.2(tau, G.array, dist)
}
\arguments{
\item{tau}{array of delay times (ms)}

\item{G.array}{array of exponent decay times (1/ms)}

\item{dist}{weigth of the decay times}
}
\value{
an array for every tau value
}
\description{
g.2 correlation from delay times, exponents and a weight distribution
}
\details{
calculate the correlation function for DLS
based on the delay times and an array for
the delay times with their weights
ideally the weights have a sum of 1

This is a code snippet used in the analysis.sizes for fitting,
but extracted to simulate the autocorrelation functions for testing
}
