% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{MSD}
\alias{MSD}
\title{MSD: mean square displacement from autocorrelation}
\usage{
MSD(a, N = 3)
}
\arguments{
\item{a}{a DLS list object. A list containing 'correlation' and 'q'
elements at least. The correlation element is a 2D array within
first column the time, the others are the normalized correlation
functions exp(-MSD/(2N) q^2).}

\item{N}{integer, the dimensionality, it should be 3 for DLS because we
detect the diffusion in 3D.}
}
\value{
an array with the same dimensions as the autocorrelation table was,
         containing the time and the MSD values
}
\description{
MSD: mean square displacement from autocorrelation
}
\details{
Calculate he mean squared displacement from the intensity autocorrelation
function. The calculation is based on the Langevin equation model of
translational diffusion under the influence of thermal forces, that is
a Brownian motion.
The model results in the autocorrelation having the form
of exp(-MSD/(2N) q^2), thus estimate: -log(correlation)/q**2
if q is in 1/micron, we get it in micron**2
Kills NaN and infinite elements to -1

A key issue of this method is the strong limit close to 1 and close to 0,
where further diffusion cannot be resolved.
}
