% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{plot.corr.series}
\alias{plot.corr.series}
\title{summary plot for several correlation data sets}
\usage{
\method{plot}{corr.series}(
  namelst,
  norm = TRUE,
  norm.range = c(4:8),
  legends = NULL,
  envir = .GlobalEnv,
  ...
)
}
\arguments{
\item{namelst}{an array of variable names}

\item{norm}{bool if true, normalize the correlation functions to 1}

\item{norm.range}{array  which points to use for normalization}

\item{legends}{text array of legends}

\item{envir}{the environment where to search for the variables}

\item{...}{all further parameters are passed to plot}
}
\value{
nothing
}
\description{
summary plot for several correlation data sets
}
\details{
Pick a list of variale names,
and plot the correlation function of each, making time the legend
Because of the list of variables, plot only the first correlation channel.

For example having experiments loaded as names
var.001, var.002, var.003...
you can use ls(pattern='^var\\.') for namelst.
}
