% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{plot.dist.series}
\alias{plot.dist.series}
\title{take a list of variable names for fits made using analysis.sizes,
and plot the size distribution from them}
\usage{
\method{plot}{dist.series}(namelst, Rh.array.refined = NULL, legends = NULL, envir = .GlobalEnv, ...)
}
\arguments{
\item{namelst}{an array of variable names containing fits}

\item{Rh.array.refined}{an array of radii to be used (optional)}

\item{legends}{array of strings to be used as legends}

\item{envir}{the environment to search for the variables}

\item{...}{extra plot parameters}
}
\value{
nothing
}
\description{
take a list of variable names for fits made using analysis.sizes,
and plot the size distribution from them
}
