% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{plot.DLS.dist}
\alias{plot.DLS.dist}
\title{plot the size distribution of a analysis.sizes fit}
\usage{
\method{plot}{DLS.dist}(
  ft,
  which = "I",
  smooth = TRUE,
  add = FALSE,
  xlab = "hydrodynamic radius, nm",
  ylab = "intensity weight",
  type = "p",
  log = "x",
  main = "",
  col = 1,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{ft}{a fit list from a size analysis}

\item{which}{a character I, N or V for the type of distribution}

\item{smooth}{Boolean,   also plot a smoothened line across
use smooth.dist above}

\item{add}{Boolean,   add to the plot, not a new plot}

\item{xlab}{text, passed to plot}

\item{ylab}{text, passed to plot}

\item{type}{character, passed to plot}

\item{log}{text passed to plot, default is 'x'}

\item{main}{text, title of plot, passed to plot}

\item{col}{text or integer, color of the plot}

\item{xlim}{array of 2, passed to plot}

\item{ylim}{array of 2, passed to plot}

\item{...}{all other parameters passed to smooth.dist()}
}
\value{
invisible return the smoothed distribution
}
\description{
plot the size distribution of a analysis.sizes fit
}
\details{
simple plot of the size distribution
}
