% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{plot.DLS}
\alias{plot.DLS}
\title{Plotting DLS data, that is correlation and its fit}
\usage{
\method{plot}{DLS}(
  dls,
  fit = NULL,
  ch = c("ch1", "ch2"),
  dual = TRUE,
  log = "x",
  xlim = NULL,
  ...
)
}
\arguments{
\item{dls}{a DLS imported data list}

\item{fit}{a cumulant fit list, not a fit object! (no predict() method!)}

\item{ch}{character array    which channels to plot}

\item{dual}{Boolean  if true, plot the residuals on the bottom}

\item{log}{text    which axis should be a log plot}

\item{xlim}{text    title of x-axis, passed to plot()}

\item{...}{passed to plot()}
}
\value{
nothing
}
\description{
Plotting DLS data, that is correlation and its fit
}
\details{
Quick plotting of the autocorrelation in semilog format as it is common
if fit is provided for the whole tau range, add the prediction of the fit
}
