% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ALV.R
\name{read.ALV}
\alias{read.ALV}
\title{read ALV correlation files}
\usage{
read.ALV(filename)
}
\arguments{
\item{filename}{text, name of the file to read (with path)}
}
\value{
a list containing all information, also adding parameters such as
         q (scattering vector), theta.rad the scattering angle in radians
}
\description{
read ALV correlation files
}
\details{
Read the ASC file output of the ALV DLS system, extract
metadata related to the experiment and the autocorrelation
tables.
Important metadata include date, time, sample name, temperature,
medium refractive index, viscosity, scattering angle

The file has a relatively fixed structure, which is read out in as text
then processed according to a predefined schema (a set of fields expected)
}
