% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CSV-data.R
\name{read.DLS.zetasizer}
\alias{read.DLS.zetasizer}
\title{read a csv file containing the correlation data from Zetasizer}
\usage{
read.DLS.zetasizer(
  filename,
  theta = 90,
  temperature = 25,
  eta = 0.89,
  n = 1.334,
  lambda = 632.8,
  microsec = TRUE,
  sep = ",",
  dec = ".",
  ch.names = c("ch1", "ch2", "ch3")
)
}
\arguments{
\item{filename}{text, the file to read}

\item{theta}{float, the scattering angle in degrees}

\item{temperature}{float, in degrees Celsius}

\item{eta}{float, the viscosity in mPas (cP)}

\item{n}{float, the refractive index of the medium}

\item{lambda}{float, the wavelength of the laser in nm}

\item{microsec}{Boolean, if the time is in microseconds
instead of milliseconds}

\item{sep}{character, the field separator}

\item{dec}{character, decimal separator, i.e. '.' or ','}

\item{ch.names}{a list of 3 names for the correlation names}
}
\value{
a list containing all the information
             time, filename, lambda, temperature,
             n, eta, theta, theta.rad, q, correlation
}
\description{
read a csv file containing the correlation data from Zetasizer
}
\details{
Read a csv file containing the correlation data from Zetasizer
and form a list for the DLS processing, adding the missing parameters.
The array is typically three X,Y columns containing the correlation
delay time as x in microseconds, and g2-1 for correlation functions,
for the steady state, transient and unfiltered versions.
However, this may be somewhat different how the export was performed.
}
