% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{seq.log}
\alias{seq.log}
\title{seq() in log-scale}
\usage{
\method{seq}{log}(start, end, len = 50)
}
\arguments{
\item{start}{double  the start value}

\item{end}{double    the end value}

\item{len}{integer   length of the resulted array}
}
\value{
an log-spaced 1D array including both ends
}
\description{
seq() in log-scale
}
\details{
logaritmic sampled number sequence
a shorthand for seq(log(start), log(end), length=len).
The resulted array is equidistant in log space.
}
