% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{smooth.dist}
\alias{smooth.dist}
\title{Calculate a smoothened version of the size distribution}
\usage{
smooth.dist(
  dls.fit,
  Rh.array = NULL,
  norm = TRUE,
  epsilon = 0.001,
  method = "natural",
  which = "I",
  ...
)
}
\arguments{
\item{dls.fit}{a fit list returned by analyse.size below}

\item{Rh.array}{the new size array where we want to have the
points. Use seq.lod to generate one}

\item{norm}{Boolean, should the distribution be normalized to
its absolute maximum?}

\item{epsilon}{float, below this value, all interpolated dist
values are set to zero}

\item{method}{string, passed to spline as method}

\item{which}{character, I,V or N for intensity, Volume or Number distribution}

\item{...}{passed to spline for further parameters}
}
\value{
a list containing x and y arrays
}
\description{
Calculate a smoothened version of the size distribution
}
\details{
The calculation is based on spline() but with logarithmic size array
}
